Collision logic implemented

+++ src/Model/PlayerObject.cpp
@@ -7,6 +7,6 @@ PlayerObject::PlayerObject () {}
 
 PlayerObject::~PlayerObject () {}
 
-float PlayerObject::getCollisionRadius() const {
-    return 25.0f;
+int PlayerObject::getCollisionRadius() const {
+    return 25;
 }
